requirements.txt
arxiv==2.1.0
openai>=2.13.0
jinja2==3.1.3
pyyaml==6.0.1
python-dotenv==1.0.1
pytest==8.0.0
pytest-mock==3.12.0
requests==2.31.0
